QPID-181 QPID-180 AbstractJMSMessage.java - updated to use getJMSHeaders JMSMapMessage.java - JMSPropertyFieldTable.java - Moved functionality of setting and retrieving a JMS property. Now shared by the Headers and MapMessageTest.java MapMessageTest.java - Updated the exceptions that are caught as all methods should throw a JMSException i.e. MessageFormatException TextMessageTest.java - Added tests for the Message Properties common/pom.xml - Added JMS dependency for the JMSPropertyFieldTable.java and associated tests EncodingUtils.java - changed comments and changed getencodedCharLength return to an int PropertyFieldTable.java - Cleaned up and now uses enum for prefixs. Created comprehensive test of both PropertyFieldTable classes. AMQPInvalidClassException.java - created to throw a runtime exception when trying to add a non-primative value. Forcing clients to handle this would break the Map usage of the FieldTable.
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@487481 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/common/src/main/java/org/apache/qpid/framing/EncodingUtils.java
@@ -461,7 +461,7 @@ public class EncodingUtils
 
     //**** new methods
 
-    // BOOLEAN_PROPERTY_PREFIX
+    // AMQP_BOOLEAN_PROPERTY_PREFIX
 
     public static void writeBoolean(ByteBuffer buffer, Boolean aBoolean)
     {
@@ -479,7 +479,7 @@ public class EncodingUtils
         return 1;
     }
 
-    // BYTE_PROPERTY_PREFIX
+    // AMQP_BYTE_PROPERTY_PREFIX
     public static void writeByte(ByteBuffer buffer, Byte aByte)
     {
         buffer.put(aByte);
@@ -496,7 +496,7 @@ public class EncodingUtils
     }
 
 
-    // SHORT_PROPERTY_PREFIX
+    // AMQP_SHORT_PROPERTY_PREFIX
     public static void writeShort(ByteBuffer buffer, Short aShort)
     {
         buffer.putShort(aShort);
@@ -528,7 +528,7 @@ public class EncodingUtils
         return 4;
     }
 
-    // LONG_PROPERTY_PREFIX
+    // AMQP_LONG_PROPERTY_PREFIX
     public static void writeLong(ByteBuffer buffer, Long aLong)
     {
         buffer.putLong(aLong);
@@ -610,7 +610,7 @@ public class EncodingUtils
     }
 
     //CHAR_PROPERTY
-    public static long encodedCharLength()
+    public static int encodedCharLength()
     {
         return encodedByteLength();
     }
